refactor(core): move type validation utilities to internal utils
